+++ R/createEllipseGate.R
@@ -40,7 +40,7 @@ createEllipseGate <- function(experimentId, xChannel, yChannel, name,
                               gid = generateId(),
                               parentPopulationId = NULL, parentPopulation = NULL,
                               tailoredPerFile = FALSE, fcsFileId = NULL, fcsFile = NULL,
-                              locked = FALSE, createPopulation = TRUE) {
+                              locked = FALSE, createPopulation = is.null(fcsFileId)) {
   body <- list(
     model = list(
       locked = jsonlite::unbox(locked),